control/requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
grpcio==1.69.0
grpcio-tools==1.69.0
itsdangerous==2.2.0
Jinja2==3.1.5
kafka-python-ng==2.2.3
MarkupSafe==3.0.2
protobuf==5.29.3
waitress==3.0.2
Werkzeug==3.1.3
requests==2.32.3
